MOON-283: Rename compile-sass to compile-css as it will include postcss as well
